Switch from travis to Github actions (#59)
Switch from Travis to Github-Actions

+++ src/metrics_classification.rs
@@ -571,7 +571,7 @@ mod tests {
 
         let x = predicted.confusion_matrix(ground_truth);
 
-        abs_diff_eq!(x.accuracy(), 5.0 / 6.0 as f32);
+        abs_diff_eq!(x.accuracy(), 5.0 / 6.0_f32);
         abs_diff_eq!(
             x.mcc(),
             (2. * 3. - 1. * 0.) / (2.0f32 * 3. * 3. * 4.).sqrt() as f32
